// RetailStation.Angular/src/app/components/Admin/models/Operation/PromotionModel.ts
import { SupplierItemModel } from "src/app/components/Main/models/SupplierItemModel";

export interface PromotionModel extends SupplierItemModel {
    promotionId: number | null;
    itemId: number | null;
    title: string;
    description: string | null;
    image: File | null;
    imageURL: string;
    startDate: string;
    endDate: string;
    isActive: boolean;
    offerPrice?: number;
    minQty?: number;
    maxQty?: number;
    totalCount?: number;
}